GetSequence should use size_t instead of int (#537)
Essentially, `GetSequence` uses `std::span::subspan` which uses
`size_t`. This PR fixes several warnings caused by implicit conversion
between `size_t` and `int`, which is one step forward towards enableing
`/WX`.

Internal work item: https://task.ms/aii/34991

+++ src/generators.h
@@ -130,7 +130,7 @@ struct Generator {
   void ComputeLogits();
   void GenerateNextToken();
 
-  RoamingArray<int32_t> GetSequence(int index) const;
+  RoamingArray<int32_t> GetSequence(size_t index) const;
 
   std::shared_ptr<const Model> model_;
   std::unique_ptr<State> state_;
